move shared types folder
