--- src/shared/types/providers.types.ts
@@ -1,1 +1,0 @@
-export type Providers = 'google'
