+++ src/types/locale.types.ts
